config.py
import os
from base64 import b64decode
from dotenv import load_dotenv
load_dotenv()

API_HASH = os.getenv("API_HASH")
API_ID = int(os.getenv("API_ID"))
MONGO_URI = os.getenv("MONGO_URI")
SESSION = os.getenv("SESSION")
PREFIX = os.getenv("PREFIX", ".")
LOGO_STUF = os.getenv("LOGO_STUF", "https://telegra.ph/file/1c6fa96c58dcb10dcf408.jpg")
LOG_CHAT = int(os.getenv("LOG_CHAT", "0"))
HEROKU_API = os.getenv("HEROKU_API", None)
HEROKU_APP_NAME = os.getenv("HEROKU_APP_NAME", None)
GIT_TOKEN = os.getenv("GIT_TOKEN", "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")
PM_LOGO = os.getenv("PM_LOGO", "https://telegra.ph/file/1c6fa96c58dcb10dcf408.jpg")
