requirements.txt
pyrogram==1.4.16
tgcrypto
aiohttp==3.8.1
aiofiles
pymongo
motor
dnspython
requests
asyncio
pillow
GitPython
heroku3
python-dotenv
ffmpeg-python
telegraph==1.4.1
numpy
googletrans==4.0.0-rc1
google==3.0.0
google-search==1.1.1
googlesearch-python== 1.0.1
gTTS
requests
wikipedia
bitlyshortener==0.3.5
datetime
speedtest-cli
wget
pyshorteners
psutil
dnspython
removebg
pytgcalls==3.0.0.dev22
youtube-search-python
yt-dlp==2021.11.10.1
prettytable
bs4
pytube
